modify for compat

+++ src/freebsd/machine/platform/x86/elf.h
@@ -33,9 +33,9 @@
 
 #if defined(__i386__) || defined(_MACHINE_ELF_WANT_32BIT)
 
-/*
- * ELF definitions for the i386 architecture.
- */
+ /*
+  * ELF definitions for the i386 architecture.
+  */
 
 #include <sys/elf32.h>	/* Definitions common to all 32 bit architectures. */
 #if defined(__ELF_WORD_SIZE) && __ELF_WORD_SIZE == 64
@@ -52,18 +52,18 @@
 
 #define	ELF_MACHINE_OK(x) ((x) == EM_386 || (x) == EM_486)
 
-/*
- * Auxiliary vector entries for passing information to the interpreter.
- *
- * The i386 supplement to the SVR4 ABI specification names this "auxv_t",
- * but POSIX lays claim to all symbols ending with "_t".
- */
+  /*
+   * Auxiliary vector entries for passing information to the interpreter.
+   *
+   * The i386 supplement to the SVR4 ABI specification names this "auxv_t",
+   * but POSIX lays claim to all symbols ending with "_t".
+   */
 
 typedef struct {	/* Auxiliary vector entry on initial stack */
 	int	a_type;			/* Entry type. */
 	union {
-		long	a_val;		/* Integer value. */
-		void	*a_ptr;		/* Address. */
+		long long	a_val;		/* Integer value. */
+		void* a_ptr;		/* Address. */
 		void	(*a_fcn)(void);	/* Function pointer (not used). */
 	} a_un;
 } Elf32_Auxinfo;
@@ -83,7 +83,7 @@ __ElfType(Auxinfo);
 
 #define	R_386_COUNT	38	/* Count of defined relocation types. */
 
-/* Define "machine" characteristics */
+ /* Define "machine" characteristics */
 #define	ELF_TARG_CLASS	ELFCLASS32
 #define	ELF_TARG_DATA	ELFDATA2LSB
 #define	ELF_TARG_MACH	EM_386
@@ -93,9 +93,9 @@ __ElfType(Auxinfo);
 
 #elif defined(__amd64__)
 
-/*
- * ELF definitions for the AMD64 architecture.
- */
+ /*
+  * ELF definitions for the AMD64 architecture.
+  */
 
 #ifndef __ELF_WORD_SIZE
 #define	__ELF_WORD_SIZE	64	/* Used by <sys/elf_generic.h> */
@@ -109,12 +109,12 @@ __ElfType(Auxinfo);
 
 #define	ELF_MACHINE_OK(x) ((x) == EM_X86_64)
 
-/*
- * Auxiliary vector entries for passing information to the interpreter.
- *
- * The i386 supplement to the SVR4 ABI specification names this "auxv_t",
- * but POSIX lays claim to all symbols ending with "_t".
- */
+  /*
+   * Auxiliary vector entries for passing information to the interpreter.
+   *
+   * The i386 supplement to the SVR4 ABI specification names this "auxv_t",
+   * but POSIX lays claim to all symbols ending with "_t".
+   */
 typedef struct {	/* Auxiliary vector entry on initial stack */
 	int	a_type;			/* Entry type. */
 	union {
@@ -123,10 +123,10 @@ typedef struct {	/* Auxiliary vector entry on initial stack */
 } Elf32_Auxinfo;
 
 typedef struct {	/* Auxiliary vector entry on initial stack */
-	long	a_type;			/* Entry type. */
+	long long	a_type;			/* Entry type. */
 	union {
-		long	a_val;		/* Integer value. */
-		void	*a_ptr;		/* Address. */
+		long long	a_val;		/* Integer value. */
+		void* a_ptr;		/* Address. */
 		void	(*a_fcn)(void);	/* Function pointer (not used). */
 	} a_un;
 } Elf64_Auxinfo;
@@ -139,7 +139,7 @@ __ElfType(Auxinfo);
 
 #define	R_X86_64_COUNT	24	/* Count of defined relocation types. */
 
-/* Define "machine" characteristics */
+ /* Define "machine" characteristics */
 #if __ELF_WORD_SIZE == 32
 #define ELF_TARG_CLASS  ELFCLASS32
 #else
