+++ src/freebsd/machine/platform/x86/x86_var.h
@@ -32,12 +32,12 @@
 #ifndef _X86_X86_VAR_H_
 #define	_X86_X86_VAR_H_
 
-/*
- * Miscellaneous machine-dependent declarations.
- */
+ /*
+  * Miscellaneous machine-dependent declarations.
+  */
 
-extern	long	Maxmem;
-extern	u_int	basemem;
+extern	u_long	Maxmem;
+extern	u_long	basemem;
 extern	int	busdma_swi_pending;
 extern	u_int	cpu_exthigh;
 extern	u_int	cpu_feature;
@@ -112,8 +112,8 @@ struct	minidumpstate;
  */
 typedef void alias_for_inthand_t(void);
 
-bool	acpi_get_fadt_bootflags(uint16_t *flagsp);
-void	*alloc_fpusave(int flags);
+bool	acpi_get_fadt_bootflags(uint16_t* flagsp);
+void* alloc_fpusave(int flags);
 void	busdma_swi(void);
 u_int	cpu_auxmsr(void);
 vm_paddr_t cpu_getmaxphyaddr(void);
@@ -123,7 +123,7 @@ void	cpu_setregs(void);
 int	dbreg_set_watchpoint(vm_offset_t addr, vm_size_t size, int access);
 int	dbreg_clr_watchpoint(vm_offset_t addr, vm_size_t size);
 void	dbreg_list_watchpoints(void);
-void	x86_clear_dbregs(struct pcb *pcb);
+void	x86_clear_dbregs(struct pcb* pcb);
 bool	disable_wp(void);
 void	restore_wp(bool old_wp);
 void	finishidentcpu(void);
@@ -134,7 +134,7 @@ void	identify_hypervisor(void);
 void	initializecpu(void);
 void	initializecpucache(void);
 bool	fix_cpuid(void);
-void	fillw(int /*u_short*/ pat, void *base, size_t cnt);
+void	fillw(int /*u_short*/ pat, void* base, size_t cnt);
 int	is_physical_memory(vm_paddr_t addr);
 int	isa_nmi(int cd);
 void	handle_ibrs_entry(void);
@@ -144,15 +144,15 @@ void	hw_mds_recalculate(void);
 void	hw_ssb_recalculate(bool all_cpus);
 void	x86_taa_recalculate(void);
 void	x86_rngds_mitg_recalculate(bool all_cpus);
-void	nmi_call_kdb(u_int cpu, u_int type, struct trapframe *frame);
-void	nmi_call_kdb_smp(u_int type, struct trapframe *frame);
-void	nmi_handle_intr(u_int type, struct trapframe *frame);
-void	pagecopy(void *from, void *to);
+void	nmi_call_kdb(u_int cpu, u_int type, struct trapframe* frame);
+void	nmi_call_kdb_smp(u_int type, struct trapframe* frame);
+void	nmi_handle_intr(u_int type, struct trapframe* frame);
+void	pagecopy(void* from, void* to);
 void	printcpuinfo(void);
 int	pti_get_default(void);
 int	user_dbreg_trap(register_t dr6);
-int	cpu_minidumpsys(struct dumperinfo *, const struct minidumpstate *);
-struct pcb *get_pcb_td(struct thread *td);
+int	cpu_minidumpsys(struct dumperinfo*, const struct minidumpstate*);
+struct pcb* get_pcb_td(struct thread* td);
 uint64_t rdtsc_ordered(void);
 
 /*
@@ -163,9 +163,9 @@ uint64_t rdtsc_ordered(void);
 #define	MSR_OP_WRITE		0x00000003
 #define	MSR_OP_READ		0x00000004
 
-/*
- * Where and which execution mode
- */
+ /*
+  * Where and which execution mode
+  */
 #define	MSR_OP_LOCAL		0x10000000
 #define	MSR_OP_SCHED_ALL	0x20000000
 #define	MSR_OP_SCHED_ONE	0x30000000
@@ -173,6 +173,6 @@ uint64_t rdtsc_ordered(void);
 #define	MSR_OP_RENDEZVOUS_ONE	0x50000000
 #define	MSR_OP_CPUID(id)	((id) << 8)
 
-void x86_msr_op(u_int msr, u_int op, uint64_t arg1, uint64_t *res);
+void x86_msr_op(u_int msr, u_int op, uint64_t arg1, uint64_t* res);
 
 #endif
